+++ src/embedding/embedding.constants.ts
@@ -1,0 +1,1 @@
+export const BGE_M3_VECTOR_SIZE = 1024;
